Moab testdata - create base setup for test data creation (#836)
* ECEEPLT-5247:Q2O Orders for UK and EMEA

* ECEEPLT-5247:Q2O Orders for UK and EMEA

* ECEEPLT-5247:Q2O Orders for UK and EMEA

* ECEEPLT-5247:Q2O Orders for UK and EMEA

* ECEEPLT-5247:Q2O Orders for UK and EMEA

* ECEEPLT-5247:Q2O Orders for UK and EMEA

* ECEEPLT-5247:Q2O Orders for UK and EMEA

* ECEEPLT-5247:Q2O Orders for UK and EMEA

* Add base test to create orders

* ECEEPLT-5330:MOAB : POC to test to create order

* ECEEPLT-5330:MOAB : POC to test to create order

* ECEEPLT-5330:MOAB : POC to test to create order

* ECEEPLT-5330:MOAB : POC to test to create order

* ECEEPLT-5330:MOAB : POC to test to create order

* update purchase email id

* Update ece/src/main/java/com/autodesk/ece/bic/testsuites/BICOrderCreation.java


* Update ece/src/main/java/com/autodesk/ece/bic/testsuites/BICOrderCreation.java


* Updated Jenkins file

* Updated Jenkins file

* update Jenkins file and modify scripts

* update address

* update

* fix typo

* udpate
:

* udpate
:

* udpate:

* updated tenant param

* commnt to not to push to project78

* updated test in jenkins

* update sales org & sku

* update sales org & sku

* merge conflicts

* merge conflicts

* merge conflicts

* merge conflicts

* merge conflicts

* Update ece/src/main/java/com/autodesk/ece/bic/testsuites/BICOrderCreation.java


* merge conflicts

+++ ece/src/main/java/com/autodesk/ece/testbase/DatastoreClient.java
@@ -9,6 +9,9 @@ import io.restassured.path.json.JsonPath;
 import io.restassured.response.Response;
 import java.io.UnsupportedEncodingException;
 import java.lang.reflect.Field;
+import java.math.BigInteger;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URLEncoder;
 import java.util.LinkedList;
 import java.util.List;
@@ -89,7 +92,7 @@ public class DatastoreClient {
     @Builder.Default
     String tenant = TENANT;
     String emailId;
-    Integer orderNumber;
+    BigInteger orderNumber;
     String quoteId;
     String paymentType;
     String locale;
@@ -103,7 +106,7 @@ public class DatastoreClient {
     String createdAt;
     String status;
 
-    OrderData(String name, String tenant, String environment, String emailId, Integer orderNumber, String quoteId,
+    OrderData(String name, String tenant, String environment, String emailId, BigInteger orderNumber, String quoteId,
         String paymentType, String locale, String address) {
       super(name, tenant, environment, emailId, orderNumber, quoteId, paymentType, locale, address);
     }
